Disclaimin’
It’s so funny that when I wanted to commit this off the github website it decided it was a good idea to autogenerate the commit message and description using copilot. Like, dear “AI,” if you are so intelligent, then maybe *take the hint and lay off*?

Anyway I deleted that crap and wrote this myself. Like everything else I submit to this website.